1. Converted Home Activity to Home Fragment 2.Added onCreateOptions 3.Updated home_activity.xml
