+++ app/src/main/java/com/nativework/covid19vaccinetracker/ui/appointment/AppointmentFragment.kt
@@ -2,14 +2,14 @@ package com.nativework.covid19vaccinetracker.ui.appointment
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import android.widget.Toast
-import androidx.core.text.isDigitsOnly
 import androidx.lifecycle.ViewModelProvider
+import com.nativework.covid19vaccinetracker.R
+import com.nativework.covid19vaccinetracker.base.BaseApp
 import com.nativework.covid19vaccinetracker.base.BaseFragment
 import com.nativework.covid19vaccinetracker.databinding.FragmentAppointmentBinding
+import com.nativework.covid19vaccinetracker.ui.HomeFragment
 import com.nativework.covid19vaccinetracker.ui.center.CenterActivity
 import com.nativework.covid19vaccinetracker.utils.Constants
 import java.text.SimpleDateFormat
@@ -21,7 +21,7 @@ class AppointmentFragment : BaseFragment() {
     private val binding get() = _binding!!
     private var viewModel: AppointmentViewModel? = null
     private var dateSelected: String? = null
-    private var pincode:Int? = 0
+    private var pincode: Int? = 0
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -29,6 +29,7 @@ class AppointmentFragment : BaseFragment() {
     ): View {
         viewModel = ViewModelProvider(this).get(AppointmentViewModel::class.java)
         _binding = FragmentAppointmentBinding.inflate(inflater, container, false)
+        setHasOptionsMenu(true)
         return binding.root
     }
 
@@ -37,8 +38,8 @@ class AppointmentFragment : BaseFragment() {
         _binding = null
     }
 
-    companion object{
-        fun newInstance(): AppointmentFragment{
+    companion object {
+        fun newInstance(): AppointmentFragment {
             val args = Bundle()
             val fragment = AppointmentFragment()
             fragment.arguments = args
@@ -49,6 +50,7 @@ class AppointmentFragment : BaseFragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        (activity as BaseApp).setToolbarTitle("Search By Pin Code")
         setListeners()
         initData()
         setObservers()
@@ -64,15 +66,13 @@ class AppointmentFragment : BaseFragment() {
             dateSelected = sdf.format(calendar.time)
         }
 
-        binding.btnSearch.setOnClickListener {
+        binding.edtPincode.setOnClickListener {
             pincode = getPincode()
         }
 
-        binding.btnSearch.setOnClickListener {
-            if (dateSelected.isNullOrEmpty() || pincode == 0) {
+        binding.btnSearchByPin.setOnClickListener {
+            if (!dateSelected.isNullOrEmpty() && pincode == 0) {
                 dateSelected = getSelectedDate()
-                pincode = getPincode()
-            }else{
                 pincode = getPincode()
                 dateSelected?.let { it1 -> pincode?.let { it2 -> getAppointment(it2, it1) } }
             }
@@ -81,11 +81,10 @@ class AppointmentFragment : BaseFragment() {
     }
 
     private fun getAppointment(i: Int, date: String) {
-        println("-----  $i  and  $date")
         viewModel?.getAppointmentByPincodeAndDate(i, date)
     }
 
-    private fun initData(){
+    private fun initData() {
         val calendar = Calendar.getInstance()
         calendar.apply {
             set(Calendar.HOUR_OF_DAY, Calendar.getInstance().getActualMinimum(Calendar.HOUR_OF_DAY))
@@ -94,11 +93,11 @@ class AppointmentFragment : BaseFragment() {
         binding.calendarView.minDate = date
     }
 
-    private fun  getPincode(): Int {
-        val pin :String = binding.edtPincode.text.trim().toString()
-        if (!pin.isNullOrEmpty()){
+    private fun getPincode(): Int {
+        val pin: String = binding.edtPincode.text.trim().toString()
+        if (!pin.isNullOrEmpty()) {
             return pin.toInt()
-        }else{
+        } else {
             Toast.makeText(context, "Enter pin code", Toast.LENGTH_SHORT).show()
         }
         return 0
@@ -132,5 +131,12 @@ class AppointmentFragment : BaseFragment() {
                 Toast.makeText(context, "Empty/No Schedule found", Toast.LENGTH_LONG).show()
             }
         })
+    }
+
+    override fun onPrepareOptionsMenu(menu: Menu) {
+        val menuPin = menu.findItem(R.id.menu_searchByPin)
+        //val menuDistrict = menu.findItem(R.id.menu_searchByDistrict)
+        menuPin.isEnabled = false
+        return
     }
 }
