+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,8 @@
     <string name="app_name">VaccineTracker</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="enter_pincode">Enter Pincode</string>
+    <string name="search">Search</string>
+    <string name="please_search_the_available_centers_for_vaccine_by_pin_code">Please search the available centers for vaccine by Pin code</string>
+    <string name="search_by_district">Search by District</string>
 </resources>
